// fof8-ui/src/Enums/Attributes.ts
export enum Attribute {
    BumpandRunDefense = "BumpandRunDefense",
    Endurance = "Endurance",
    Intercepting = "Intercepting",
    MantoManDefense = "MantoManDefense",
    PlayDiagnosis = "PlayDiagnosis",
    PunishingHitter = "PunishingHitter",
    RunDefense = "RunDefense",
    ZoneDefense = "ZoneDefense"
}